Fixed crash that occurred when searching for some repositories
This occurred because the GitHub API returns 3 types: 'blobs',
'trees', and 'commits'. Since 'blobs' and 'trees' are displayed as
files and folders, the FileExplorer component didn't know how to display
nodes with the type 'commit'. For now, nodes with type 'commit' are ignored.
Maybe this'll be a feature in the future???